+++ CMakeLists.txt
@@ -1,3 +1,3 @@
 cmake_minimum_required(VERSION 3.22.1)
 project(Simulador_de_trafego)
-add_executable(Simulador simulacao.cpp)
+add_executable(Simulador_Final main.cpp Veiculo.cpp Rodovia.cpp Simulador.cpp)
