+++ main.cpp
@@ -1,0 +1,15 @@
+#include "Simulador.h"
+#include "Veiculo.h"
+#include "Rodovia.h"
+
+using namespace std;
+using namespace N;
+
+int main() {
+    Simulacao S1(3);
+
+    S1.simulacao();
+    
+    
+    return 0;
+}
